requirements.txt
asgiref==3.8.1
Django==5.1
sqlparse==0.5.1
tzdata==2024.1
django-extensions==3.2.3
unidecode==1.3.8
markdown==3.7
Pillow==10.4.0
django-debug-toolbar==4.4.6